etl_app/requirements.txt
psycopg2-binary
elasticsearch
elasticsearch_dsl
pydantic
